added git installation instructions
